Framework documentation and cleanup (WIP).
