--- src/Enecuum/API.hs
@@ -1,5 +1,0 @@
-module Enecuum.API 
-  ( module X
-  ) where
-
-import Enecuum.API.RPC as X
