+++ src/Enecuum/Core/Language.hs
@@ -7,6 +7,7 @@ import Enecuum.Prelude
 
 import Enecuum.Core.Logger.Language as X
 
+-- | Core effects represent general effects every scenario can make.
 type CoreEffects =
   '[ LoggerL
    , SIO
