+++ src/Enecuum/Core/Types/Logger.hs
@@ -1,4 +1,5 @@
 module Enecuum.Core.Types.Logger where
 
 
+-- | Logging level.
 data LogLevel = Info
