+++ src/Enecuum/Framework/Domain.hs
@@ -6,5 +6,4 @@ where
 import           Enecuum.Framework.Domain.NetworkModel as X
 import           Enecuum.Framework.Domain.Networking as X
 import           Enecuum.Framework.Domain.Node as X
-import           Enecuum.Framework.Domain.NodeDefinition as X
 import           Enecuum.Framework.Domain.Error as X
