+++ src/Enecuum/Framework/Domain/NetworkModel.hs
@@ -5,7 +5,8 @@ module Enecuum.Framework.Domain.NetworkModel where
 
 import           Enecuum.Prelude
 
--- This is a raw vision of networking model. Will be updated later.
+-- This is a raw vision of networking model.
+-- There is a probability we don't need this model at all.
 
 data NetworkConfig = NetworkConfig
 
@@ -14,6 +15,9 @@ data DNSResponse = DNSResponse
 
 -- TODO: needs more type safety to prevent passing MulticastRequest to broadcast (and vice versa).
 -- Use TF
+
+-- These types are just dummy.
+
 data NetworkRequest = MulticastRequest
   { dummy :: Text
 
@@ -23,6 +27,8 @@ data NetworkResponse = NetworkResponse
   { dummy :: Text
   }
 
+-- | Declares how to convert to and from low-level network requests and responses.
+-- This is a draft.
 class NetworkMethod cfg req resp | req -> resp, resp -> req where
     toNetworkRequest :: cfg -> req -> NetworkRequest
     fromNetworkResponse :: cfg -> NetworkResponse -> Maybe resp
