--- src/Enecuum/Framework/Domain/NodeDefinition.hs
@@ -1,7 +1,0 @@
-module Enecuum.Framework.Domain.NodeDefinition where
-
-import Enecuum.Framework.Domain.Node (NodeID)
-
-data NodeDef = NodeDef
-  { nodeID :: NodeID
-  }
