+++ test/Enecuum/Core/Testing/Runtime/Logger/Impl.hs
@@ -2,18 +2,21 @@ module Enecuum.Core.Testing.Runtime.Logger.Impl where
 
 import Enecuum.Prelude
 
-import           Eff (Eff, Member, handleRelay, runM, send, raise, replaceRelay)
+import           Eff (Eff, handleRelay)
 
-import qualified Enecuum.Core.Types                         as D
 import qualified Enecuum.Core.Language                      as L
 import qualified Enecuum.Core.Testing.Runtime.Lens          as RLens
 import           Enecuum.Core.Testing.Runtime.Types
+
+
+-- | Interprets a LoggerL language.
+-- Just pushes the messages into the concurrent list-like storage.
 
 interpretLoggerL
   :: LoggerRuntime
   -> L.LoggerL a
   -> Eff '[SIO, Exc SomeException] a
-interpretLoggerL rt (L.LogMessage level msg) =
+interpretLoggerL rt (L.LogMessage _ msg) =
   safeIO $ atomically $ modifyTVar (rt ^. RLens.messages) (msg :)
 
 runLoggerL
